+++ library/LocalizationObject.mjs
@@ -2,5 +2,5 @@
 /** @typedef {import("./Texts.mjs").Texts} Texts */
 
 /**
- * @typedef {{"additional-system-languages"?: string[] | null, default?: boolean | null, direction?: string | null, label?: string | ((localization: Localization, system_label?: string | null) => Promise<string | null>) | null, language: string, texts?: {[key: string]: Texts} | (() => Promise<{[key: string]: Texts} | null>) | null}} LocalizationObject
+ * @typedef {{"additional-system-languages"?: string[] | null, default?: boolean | null, direction?: string | null, label?: string | ((localization: Localization, system_label?: string | null) => Promise<string | null>) | null, language: string, texts?: {[key: string]: Texts} | ((module: string) => Promise<Texts | null>) | null}} LocalizationObject
  */